weather_etl/requirements.txt
apache-airflow==2.7.1
pandas==2.1.1
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
